// src/types/dynamicSidepanel/sidebarPanels.ts
/**
 * Declare every panels name here

 */

export enum SIDEPANEL_IDS {
  ORDER_INFO = 'ORDER_INFO',
  REPORT_INFO = 'REPORT_INFO',
  DICOM_INFO = 'STUDY_INFO',
  ORDER_HISTORY = 'ORDER_HISTORY',
  ORDER_ATTACHMENT = 'FILE_ATTACH',
  VIEWER = 'EXTEND_VIEWER',
  BOOKMARK = 'BOOKMARK',
  // CONSUMABLE = 'CONSUMABLE',
  NONDICOM_CAPTURE = 'NONDICOM_CAPTURE',
}